Ben/oob python param (#284)
* add OOB and notebook showing it's use in python

fix binding

fix c++

* python: oob_score_ is accuracy

rfTree.h moving method to be same loc as rerfTree.h

* base RerF treeStats dict type -> float

update notebook

move docstring rerfClassifier consistent w/ inputs

* update/re-ran iris perfect test

+++ packedForest/src/forestTypes/basicForests/rerf/fpRerFBase.h
@@ -67,7 +67,7 @@ namespace fp {
 				;
 			}
 
-			inline std::map<std::string, int> calcTreeStats(){
+			inline std::map<std::string, float> calcTreeStats(){
 				int maxDepth=0;
 				int totalLeafNodes=0;
 				int totalLeafDepth=0;
@@ -81,16 +81,16 @@ namespace fp {
 					totalLeafDepth += trees[i].returnLeafDepthSum();
 				}
 
-				std::map<std::string, int> treeStats;
-				treeStats["maxDepth"] = maxDepth;
-				treeStats["totalLeafDepth"] = totalLeafDepth;
-				treeStats["totalLeafNodes"] = totalLeafNodes;
+				std::map<std::string, float> treeStats;
+				treeStats["maxDepth"] = (float) maxDepth;
+				treeStats["totalLeafDepth"] = (float) totalLeafDepth;
+				treeStats["totalLeafNodes"] = (float) totalLeafNodes;
 				treeStats["OOBaccuracy"] = reportOOB();
 				return treeStats;
 			}
 
 			inline void treeStats(){
-				std::map<std::string, int> treeStats = calcTreeStats();
+				std::map<std::string, float> treeStats = calcTreeStats();
 
 				std::cout << "max depth: " << treeStats["maxDepth"] << "\n";
 				std::cout << "avg depth: " << float(treeStats["totalLeafDepth"])/float(treeStats["totalLeafNodes"]) << "\n";
